Bug 1171: Made coeffCount uint32 instead of size_t, because there's neither a LSQFit::set(size_t) overload, nor a LSQFit::set(long)

+++ branches/CEP-BB-Bug1158-task-branch/CEP/BB/BBSKernel/include/BBSKernel/Prediffer.h
@@ -237,7 +237,7 @@ private:
 
     //# Sum of the maximal number (over all solution cells) of coefficients of
     //# each parameter.
-    size_t                              itsCoeffCount;
+    uint32                              itsCoeffCount;
     vector<MeqPExpr>                    itsParameterSelection;
     map<uint32, CellCoeffIndex>         itsCellCoeffIndices;
     //# ------------------------------------------------------------------------
